+++ changelog.md
@@ -12,6 +12,7 @@
 ## Table of Contents
 
 * [Changelog](#changelog)
+  * [Release v1.5.2](#release-v152)
   * [Release v1.5.1](#release-v151)
   * [Release v1.5.0](#release-v150)
   * [Release v1.4.1](#release-v141)
@@ -28,6 +29,13 @@
 ---
 
 ## Changelog
+
+### Release v1.5.2
+
+1. Optimize code by using passing by `reference` instead of by `value`
+2. Optional `Board_Name` in Menu. Check [option to remove board name from web page #25](https://github.com/khoih-prog/WiFiManager_NINA_Lite/issues/25)
+3. Add function `isConfigMode()` to signal system is in Config Portal mode.
+4. Update `Packages' Patches`
 
 ### Release v1.5.1